+++ src/utils/postSubscribe.js
@@ -1,16 +1,16 @@
 import 'babel-polyfill';
 import delay from 'delay';
-import * as Videos_new from '../models/Videos_new';
-import * as Push_new_videos_logs from '../models/push_new_videos_logs';
-import * as Subscribe from '../models/subscribe';
+import * as newVideos from '../models/newVideos';
+import * as pushNewVideosLogss from '../models/pushNewVideosLogs';
+import * as subscribe from '../models/subscribe';
 import FacebookOP from './facebook';
 
 const fb = new FacebookOP();
 
 
 const postSubscribe = async () => {
-  const returnArr = await Videos_new.getRandomThreeVideos();
-  const senderIDArr = await Subscribe.findSubscribeId();
+  const returnArr = await newVideos.getRandomThreeVideos();
+  const senderIDArr = await subscribe.findSubscribeId();
   const senderIDArrLength = senderIDArr.length;
 
   let successNumber = 0;
@@ -28,12 +28,12 @@ const postSubscribe = async () => {
       overOneDayNumber++;
     } else {
       failedNumber++;
-      Subscribe.updateSubscribeData(senderIDArr[idx].senderID, false);
+      subscribe.updateSubscribeData(senderIDArr[idx].senderID, false);
     }
     console.log(`需要推播人數：${senderIDArrLength} ｜ 推播成功：${successNumber} ｜ 24小時內未回覆：${overOneDayNumber} ｜ 推播失敗：${failedNumber}`);
 
     if ((idx + 1) === senderIDArrLength) {
-      Push_new_videos_logs.savePushNewVideoData(idx + 1, senderIDArrLength, successNumber, overOneDayNumber, failedNumber);
+      pushNewVideosLogss.savePushNewVideoData(idx + 1, senderIDArrLength, successNumber, overOneDayNumber, failedNumber);
     }
   }
 };
